fix : 전량 매도 종목 자산에서 제외

+++ src/main/java/com/project/flowfinserver/repository/AssetItemRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 public interface AssetItemRepository extends JpaRepository<AssetItem, Integer> {
 
-    void deleteByAccountId(Long accountId);
+    void deleteByAccountId(Integer accountId);
 
     void deleteAllByUserId(Long userId);
